pyproject.toml
[project]
name = "galley"
version = "0.1.0"
description = "OCI presales support MCP server"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "pyyaml>=6.0",
    "uvicorn",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio",
    "ruff",
    "mypy",
    "types-PyYAML",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/galley"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "src"
packages = ["galley"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.1",
    "types-pyyaml>=6.0.12.20250915",
]
